+++ ActionLogEntry.cs
@@ -1,0 +1,8 @@
+﻿namespace FileSync.Model
+{
+    public class ActionLogEntry
+    {
+        public SyncAction Action { get; set; }
+        public string RelativePath { get; set; }
+    }
+}
