+++ Enums/SyncAction.cs
@@ -1,0 +1,9 @@
+﻿namespace FileSync.Enums
+{
+    public enum SyncAction
+    {
+        Create,
+        Update,
+        Delete
+    }
+}
