--- ILogManager.cs
@@ -1,5 +1,0 @@
-﻿public interface ILogManager
-{
-  void SaveLog(SyncLog log);
-  SyncLog LoadLog();
-}
