+++ Logging/ILogManager.cs
@@ -1,0 +1,10 @@
+﻿namespace FileSync.Logging
+{
+    using FileSync.Models;
+
+    public interface ILogManager
+    {
+        void SaveLog(SyncLog log);
+        SyncLog LoadLog();
+    }
+}
