+++ Model/FileInfo.cs
