--- SyncLog.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-public class SyncLog
-{
-  public DateTime SyncTime { get; set; }
-  public List<FileInfo> Files { get; set; } = new List<FileInfo>();
-  public List<string> Actions { get; set; } = new List<string>();
-}
