+++ View/ISyncView.cs
